Move isolated vector tests out of test_PDB.py
